Total file organization refactor: grouped by components
