+++ backend/src/polydash/miners_ratings/rating_func.py
