+++ backend/src/polydash/p2p_data/model.py
@@ -1,6 +1,21 @@
 from pony import orm
+from pony.orm import PrimaryKey
 
-from polydash.db import db
+from common import GetOrInsertMixin
+from common.dashboard.db import db
+
+
+class TransactionP2P(db.Entity, GetOrInsertMixin):
+    _table_ = 'tx_summary'
+    tx_hash = orm.Required(str, index=True)
+    peer_id = orm.Required(str)
+    PrimaryKey(tx_hash, peer_id)
+    tx_first_seen = orm.Optional(int, size=64)
+
+    @classmethod
+    def get_first_by_hash(cls, tx_hash):
+        return cls.select().where(tx_hash=tx_hash).order_by(
+            cls.tx_first_seen).first()
 
 
 class BlockP2P(db.Entity):
